+++ libs/components/src/lib/components/bracket/stories/dummy-data/index.ts
@@ -1,0 +1,4 @@
+export * from './ET_DUMMY_DATA_DOUBLE_16';
+export * from './ET_DUMMY_DATA_DOUBLE_8';
+export * from './ET_DUMMY_DATA_SINGLE_16';
+export * from './FIFA_DUMMY_DATA_DOUBLE_4';