+++ libs/components/src/lib/components/overlay/components/bottom-sheet/partials/bottom-sheet-drag-handle/bottom-sheet-drag-handle.component.ts
@@ -3,13 +3,13 @@ import {
   OnChanges,
   Input,
   HostBinding,
-  Optional,
   ElementRef,
   SimpleChanges,
   HostListener,
   Component,
   ViewEncapsulation,
   ChangeDetectionStrategy,
+  inject,
 } from '@angular/core';
 import { BottomSheetService, BottomSheetSwipeHandlerService } from '../../services';
 import { BottomSheetRef, getClosestBottomSheet } from '../../utils';
@@ -29,6 +29,11 @@ import { BottomSheetRef, getClosestBottomSheet } from '../../utils';
   providers: [BottomSheetSwipeHandlerService],
 })
 export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
+  private _bottomSheetRef = inject(BottomSheetRef, { optional: true });
+  private readonly _elementRef = inject<ElementRef<HTMLElement>>(ElementRef);
+  private readonly _bottomSheetService = inject(BottomSheetService);
+  private readonly _bottomSheetSwipeHandlerService = inject(BottomSheetSwipeHandlerService);
+
   @Input('aria-label')
   ariaLabel?: string = 'Close sheet';
 
@@ -44,22 +49,15 @@ export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
 
   private _swipeHandlerId: number | null = null;
 
-  constructor(
-    @Optional() public bottomSheetRef: BottomSheetRef<unknown>,
-    private _elementRef: ElementRef<HTMLElement>,
-    private _bottomSheetService: BottomSheetService,
-    private _bottomSheetSwipeHandlerService: BottomSheetSwipeHandlerService,
-  ) {}
-
   ngOnInit() {
-    if (!this.bottomSheetRef) {
+    if (!this._bottomSheetRef) {
       const closestRef = getClosestBottomSheet(this._elementRef, this._bottomSheetService.openBottomSheets);
 
       if (!closestRef) {
         throw Error('No closest ref found');
       }
 
-      this.bottomSheetRef = closestRef;
+      this._bottomSheetRef = closestRef;
     }
   }
 
@@ -73,8 +71,12 @@ export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
 
   @HostListener('click', ['$event'])
   _onButtonClick(event: MouseEvent) {
-    this.bottomSheetRef._closeBottomSheetVia(
-      this.bottomSheetRef,
+    if (!this._bottomSheetRef) {
+      return;
+    }
+
+    this._bottomSheetRef._closeBottomSheetVia(
+      this._bottomSheetRef,
       event.screenX === 0 && event.screenY === 0 ? 'keyboard' : 'mouse',
       this.bottomSheetResult,
     );
@@ -82,9 +84,13 @@ export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
 
   @HostListener('touchstart', ['$event'])
   _onTouchStart(event: TouchEvent) {
+    if (!this._bottomSheetRef) {
+      return;
+    }
+
     this._swipeHandlerId = this._bottomSheetSwipeHandlerService.startSwipe(
       event,
-      this.bottomSheetRef._containerInstance.elementRef.nativeElement,
+      this._bottomSheetRef._containerInstance.elementRef.nativeElement,
     );
   }
 
@@ -111,7 +117,11 @@ export class BottomSheetDragHandleComponent implements OnInit, OnChanges {
     this._swipeHandlerId = null;
 
     if (shouldClose) {
-      this.bottomSheetRef._closeBottomSheetVia(this.bottomSheetRef, 'touch', this.bottomSheetResult);
+      if (!this._bottomSheetRef) {
+        return;
+      }
+
+      this._bottomSheetRef._closeBottomSheetVia(this._bottomSheetRef, 'touch', this.bottomSheetResult);
     }
   }
 }