+++ libs/components/src/lib/components/overlay/components/tooltip/utils/tooltip.util.ts
@@ -1,0 +1,6 @@
+import { TOOLTIP_CONFIG } from '../constants';
+import { TooltipConfig } from './tooltip-config';
+
+export const provideTooltipConfig = (tooltipConfig: TooltipConfig) => {
+  return { provide: TOOLTIP_CONFIG, useValue: tooltipConfig };
+};