+++ libs/components/src/lib/components/pagination/pagination.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { PaginationComponent } from './components';
+
+@NgModule({
+  imports: [PaginationComponent],
+  exports: [PaginationComponent],
+})
+export class PaginationModule {}