+++ libs/components/src/lib/components/sort/components/sort-header/public-api.ts
@@ -1,0 +1,2 @@
+export * from './sort-header.component';
+export * from './sort-header.types';