+++ libs/components/src/lib/components/sort/public-api.ts
@@ -1,0 +1,6 @@
+export * from './components/public-api';
+export * from './partials/public-api';
+export * from './services/public-api';
+export * from './types/public-api';
+
+export * from './sort.module';