+++ libs/components/src/lib/components/table/types/table-data-source.types.ts
@@ -1,0 +1,17 @@
+import { Observable, Subject } from 'rxjs';
+
+export interface TableDataSourcePageEvent {
+  pageIndex: number;
+  pageSize: number;
+  length: number;
+}
+
+export interface TableDataSourcePaginator {
+  page: Subject<TableDataSourcePageEvent>;
+  pageIndex: number;
+  initialized: Observable<void>;
+  pageSize: number;
+  length: number;
+  firstPage: () => void;
+  lastPage: () => void;
+}