+++ libs/core/src/lib/pipes/to-array/to-array.pipe.ts
@@ -1,0 +1,7 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import { toArray } from './to-array.util';
+
+@Pipe({ name: 'toArray', standalone: true })
+export class ToArrayPipe implements PipeTransform {
+  transform = toArray;
+}