+++ libs/core/src/lib/services/viewport.types.ts
@@ -1,0 +1,6 @@
+import { Breakpoint } from '../types';
+
+export type BuildMediaQueryOptions = {
+  min?: number | Breakpoint;
+  max?: number | Breakpoint;
+};