+++ libs/components/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @ethlete/components
 
+## 0.2.0-next.6
+
+### Minor Changes
+
+- [`89bf895`](https://github.com/ethlete-io/ethdk/commit/89bf89552bfaaf697edc81b558ef5e23e6508632) Thanks [@TomTomB](https://github.com/TomTomB)! - Add bracket component
+
+## 0.2.0-next.5
+
+### Patch Changes
+
+- [`a6e361e`](https://github.com/ethlete-io/ethdk/commit/a6e361e9d988a208751a4877a629fe738f902fad) Thanks [@TomTomB](https://github.com/TomTomB)! - Fix scrollable observer breaking once it switches from scrollable to not scrollable and back again
+
+## 0.2.0-next.4
+
+### Minor Changes
+
+- [`f165327`](https://github.com/ethlete-io/ethdk/commit/f1653275fe2e55c0f424d1650e174dc8900cb7d4) Thanks [@TomTomB](https://github.com/TomTomB)! - Add accordion component
+
+## 0.2.0-next.3
+
+### Minor Changes
+
+- [`d2ac541`](https://github.com/ethlete-io/ethdk/commit/d2ac5415b4ff6376bdc35e5359249b59929a17b0) Thanks [@TomTomB](https://github.com/TomTomB)! - Add tab components
+
+## 0.2.0-next.2
+
+### Patch Changes
+
+- [`19eed77`](https://github.com/ethlete-io/ethdk/commit/19eed779fee863b4566b20f3f9ca6c36d66aef8f) Thanks [@TomTomB](https://github.com/TomTomB)! - Add scrollable and toggletip components
+
+## 0.2.0-next.1
+
+### Patch Changes
+
+- [#65](https://github.com/ethlete-io/ethdk/pull/65) [`1cf9a35`](https://github.com/ethlete-io/ethdk/commit/1cf9a35a39885c1054721db20ce23424c81bad74) Thanks [@TomTomB](https://github.com/TomTomB)! - Add missing exports
+
+## 0.2.0-next.0
+
+### Minor Changes
+
+- [#49](https://github.com/ethlete-io/ethdk/pull/49) [`5c95058`](https://github.com/ethlete-io/ethdk/commit/5c9505837ee3e5f2457169591acd01c79eade565) Thanks [@TomTomB](https://github.com/TomTomB)! - - Add table component
+
+  - Add pagination component
+  - Add skeleton component
+  - Add sort component
+  - Add tooltip component
+
+- [#35](https://github.com/ethlete-io/ethdk/pull/35) [`c94046b`](https://github.com/ethlete-io/ethdk/commit/c94046b64ff2986f8518bed7cc9a4f5104bd7972) Thanks [@TomTomB](https://github.com/TomTomB)! - Add dynamic overlay service
+
 ## 0.1.0
 
 ### Minor Changes